requirements.txt
PyQt6
qt-material
pillow
imageio-ffmpeg
opencv-python
send2trash
